move some hit functions outside of app
