+++ packages/utils/collision.ts
@@ -1,4 +1,4 @@
-import { Point, Polygon, Shape } from "./geometry/shape";
+import { Point, Polygon, GeometricShape } from "./geometry/shape";
 import {
   pointInEllipse,
   pointInPolygon,
@@ -12,7 +12,11 @@ import {
 } from "./geometry/geometry";
 
 // check if the given point is considered on the given shape's border
-export const isPointOnShape = (point: Point, shape: Shape, tolerance = 0) => {
+export const isPointOnShape = (
+  point: Point,
+  shape: GeometricShape,
+  tolerance = 0,
+) => {
   // get the distance from the given point to the given element
   // check if the distance is within the given epsilon range
   switch (shape.type) {
@@ -34,7 +38,7 @@ export const isPointOnShape = (point: Point, shape: Shape, tolerance = 0) => {
 };
 
 // check if the given point is considered inside the element's border
-export const isPointInShape = (point: Point, shape: Shape) => {
+export const isPointInShape = (point: Point, shape: GeometricShape) => {
   switch (shape.type) {
     case "polygon":
       return pointInPolygon(point, shape.data);
